Cleanup statusbar util, remove unused code, move some stuff around so we are exporting less API surface area.
