+++ packages/statusbar-extension/src/context.ts
