+++ packages/statusbar/src/index.ts
@@ -6,4 +6,3 @@
 export * from './statusbar';
 export * from './style/statusbar';
 export * from './components';
-export * from './util';
