--- packages/statusbar/src/util/index.ts
@@ -1,4 +1,0 @@
-export * from './settings';
-export * from './set';
-export * from './signal';
-export * from './text';
